cast score as int before sending to the server

+++ lib/pangea/models/practice_activities.dart/practice_activity_record_model.dart
@@ -169,7 +169,7 @@ class ActivityRecordResponse {
       'audio': audioBytes,
       'image': imageBytes,
       'timestamp': timestamp.toIso8601String(),
-      'score': score,
+      'score': score.toInt(),
     };
   }
